fix: resolve react-doctor error in slash-command-list and extract link utilities
- Replace useEffect state reset with direct comparison in SlashCommandList
- Extract link utility functions to markdown-editor/link-utils.ts
- Remove unused imports (createHeadingId, WINDOWS_DRIVE_PATH_PATTERN)
