+++ apps/desktop/src/components/markdown-editor.tsx
@@ -15,7 +15,6 @@ import { TextSelection } from "@tiptap/pm/state";
 import { EditorContent, useEditor } from "@tiptap/react";
 import { Markdown } from "tiptap-markdown";
 
-import { createHeadingId } from "@/core/note-navigation";
 import { getFolderRevealLabel } from "@/core/platform";
 
 import { CustomCodeBlockLowlight } from "./tiptap-extension/code-block-lowlight";
@@ -40,6 +39,20 @@ import { FindPanel } from "./find-panel";
 import { LinkPreview } from "./link-preview";
 import { ImageControls } from "./image-controls";
 import { ArrowUpIcon, OutlineIcon } from "./icons";
+import {
+  clamp,
+  collectEditorOutline,
+  DEFAULT_SELECTION_SNAPSHOT,
+  EMPTY_FIND_PANEL_STATE,
+  extractLinkAttributes,
+  getSelectedEditorText,
+  INACTIVE_TABLE_CONTROLS,
+  isExternalLink,
+  isSafeLocalLinkTarget,
+  LINK_IMAGE_PATTERN,
+  MARKDOWN_FILE_SUFFIX_PATTERN,
+  normalizeLinkTarget,
+} from "./markdown-editor/link-utils";
 
 import type {
   EditorActionDetail,
@@ -54,10 +67,6 @@ import type {
   TableControlsState,
 } from "../types/markdown-editor";
 import type { UpdateState } from "@/core/workspace";
-
-const LINK_IMAGE_PATTERN = /(!?)\[([^\]]+)\]\(([^)]+)\)$/;
-const MARKDOWN_FILE_SUFFIX_PATTERN = /\.(md|mdx|markdown)$/i;
-const WINDOWS_DRIVE_PATH_PATTERN = /^[a-z]:[\\/]/i;
 
 const getDevPreviewUpdateState = (): UpdateState | null => {
   if (!import.meta.env.DEV || typeof window === "undefined") {
@@ -131,133 +140,6 @@ const getDevPreviewUpdateState = (): UpdateState | null => {
   }
 
   return null;
-};
-
-const DEFAULT_SELECTION_SNAPSHOT: SelectionSnapshot = {
-  from: 1,
-  to: 1,
-};
-
-const EMPTY_FIND_PANEL_STATE: FindPanelState = {
-  activeIndex: -1,
-  matchCount: 0,
-};
-
-const extractLinkAttributes = (input: string) => {
-  const match = input.match(/(.+?)\s+"([^"]+)"$/);
-  if (match) {
-    return { href: match[1].trim(), title: match[2] };
-  }
-
-  return { href: input.trim(), title: undefined };
-};
-
-const normalizeLinkTarget = (value: string) => {
-  const trimmed = value.trim();
-  if (!trimmed) {
-    return "";
-  }
-
-  if (/^[a-z]+:/i.test(trimmed) && !/^(https?:\/\/|file:\/\/|glyph-local:\/\/)/i.test(trimmed)) {
-    return "";
-  }
-
-  if (/^(https?:\/\/|file:\/\/|glyph-local:\/\/)/i.test(trimmed)) {
-    return trimmed;
-  }
-
-  if (/^[\w.-]+\.[a-z]{2,}(\/.*)?$/i.test(trimmed)) {
-    return `https://${trimmed}`;
-  }
-
-  return trimmed;
-};
-
-const isSafeLocalLinkTarget = (value: string) => {
-  const trimmed = value.trim();
-  if (!trimmed) {
-    return false;
-  }
-
-  if (trimmed.startsWith("#")) {
-    return true;
-  }
-
-  if (WINDOWS_DRIVE_PATH_PATTERN.test(trimmed)) {
-    return true;
-  }
-
-  if (/^[a-z][a-z0-9+.-]*:/i.test(trimmed)) {
-    return false;
-  }
-
-  const [pathPart] = trimmed.split(/[?#]/, 1);
-  if (!pathPart || pathPart.startsWith("//")) {
-    return false;
-  }
-
-  return pathPart.includes("/") || pathPart.startsWith(".");
-};
-
-const isExternalLink = (href: string) => {
-  const trimmed = href.trim();
-  return /^https?:\/\//i.test(trimmed) || /^mailto:/i.test(trimmed);
-};
-
-const clamp = (value: number, min: number, max: number) => Math.min(Math.max(value, min), max);
-
-function collectEditorOutline(editor: Editor): EditorOutlineItem[] {
-  const items: EditorOutlineItem[] = [];
-  const counts = new Map<string, number>();
-
-  editor.state.doc.descendants((node, pos) => {
-    if (node.type.name !== "heading") {
-      return;
-    }
-
-    const level = Number(node.attrs.level ?? 0);
-    if (level < 1 || level > 4) {
-      return;
-    }
-
-    const title = node.textContent.trim();
-    if (!title) {
-      return;
-    }
-
-    const baseId = createHeadingId(title);
-    const instanceCount = counts.get(baseId) ?? 0;
-    counts.set(baseId, instanceCount + 1);
-
-    const textBeforeHeading = editor.state.doc.textBetween(0, pos, "\n");
-    const line = textBeforeHeading ? textBeforeHeading.split("\n").length + 1 : 1;
-
-    items.push({
-      id: instanceCount === 0 ? baseId : `${baseId}-${instanceCount + 1}`,
-      depth: level,
-      title,
-      line,
-      pos: pos + 1,
-    });
-  });
-
-  return items;
-}
-
-function getSelectedEditorText(editor: Editor) {
-  const { from, to } = editor.state.selection;
-  if (from === to) {
-    return "";
-  }
-
-  return editor.state.doc.textBetween(from, to, " ", " ").trim();
-}
-
-const INACTIVE_TABLE_CONTROLS: TableControlsState = {
-  active: false,
-  canDeleteRow: false,
-  canDeleteColumn: false,
-  canDeleteTable: false,
 };
 
 const runMarkdownShortcutConversion = (
