+++ apps/desktop/src/components/slash-command-list.tsx
@@ -1,4 +1,4 @@
-import { forwardRef, useCallback, useEffect, useImperativeHandle, useRef, useState } from "react";
+import { forwardRef, useCallback, useImperativeHandle, useRef, useState } from "react";
 import { formatKeysForPlatform, splitShortcutTokens } from "@/core/shortcuts";
 
 import type {
@@ -30,16 +30,21 @@ function ShortcutBadge({ shortcut }: { shortcut: string }) {
 export const SlashCommandList = forwardRef<SlashCommandListHandle, SlashCommandListProps>(
   ({ items, onSelect }, ref) => {
     const [selectedIndex, setSelectedIndex] = useState(0);
+    const prevItemsLengthRef = useRef(items.length);
     const selectedRef = useRef<HTMLButtonElement | null>(null);
     const containerRef = useRef<HTMLDivElement | null>(null);
 
-    useEffect(() => {
+    if (items.length !== prevItemsLengthRef.current) {
+      prevItemsLengthRef.current = items.length;
       setSelectedIndex(0);
-    }, [items]);
+    }
 
-    useEffect(() => {
-      selectedRef.current?.scrollIntoView({ block: "nearest" });
-    }, [selectedIndex]);
+    const handleSelectionChange = useCallback((index: number) => {
+      setSelectedIndex(index);
+      setTimeout(() => {
+        selectedRef.current?.scrollIntoView({ block: "nearest" });
+      }, 0);
+    }, []);
 
     const selectItem = useCallback(
       (index: number) => {
@@ -94,7 +99,7 @@ export const SlashCommandList = forwardRef<SlashCommandListHandle, SlashCommandL
               ref={isSelected ? selectedRef : null}
               className={`slash-row${isSelected ? " slash-row--selected" : ""}`}
               onClick={() => selectItem(index)}
-              onMouseEnter={() => setSelectedIndex(index)}
+              onMouseEnter={() => handleSelectionChange(index)}
             >
               <span className="slash-row-label">{item.title}</span>
               {item.shortcut ? <ShortcutBadge shortcut={item.shortcut} /> : null}
