+++ apps/web/public/rss.xml
@@ -277,7 +277,7 @@
     <item>
       <title>Glyph vChangelog</title>
       <link>https://glyph.falakgala.dev/changelog/</link>
-      <pubDate>Sun, 19 Apr 2026 00:00:00 GMT</pubDate>
+      <pubDate>Sat, 25 Apr 2026 00:00:00 GMT</pubDate>
       <guid>https://glyph.falakgala.dev/changelog/#vChangelog</guid>
       <description><![CDATA[All notable changes to Glyph will be documented in this file.
